fixing add appointment

+++ controllers/appointments.js
@@ -79,6 +79,22 @@ exports.addAppointment = async (req, res, next) => {
                 message: `No hospital with the id of ${req.params.hospitalId}`
             });
         }
+
+        //add user Id to req.body
+        req.body.user = req.user.id;
+        //Check for existing appointment
+        const existedAppointments = await Appointment.find({user: req.user.id});
+        //if the user is not an admin, they can create up to 3 appointment
+        if(existedAppointments.length >= 3 && req.user.role !== 'admin'){
+            return res.status(400).json({
+                success: false,
+                message: `The user with ID ${req.user.id} has already made 3 appointments`
+            });
+        }
+
+        //right now the admin can only make appointment for himself
+        //what if check for role != admin then make req.body.user = req.user.id
+
         const appointment = await Appointment.create(req.body);
         res.status(200).json({
             success: true,
@@ -97,6 +113,14 @@ exports.updateAppointment = async(req, res, next) => {
         if(!appointment){
             return res.status(404).json({success:false, message: 
                 `No appointment with id of ${req.params.id}`
+            });
+        }
+
+        //Make sure user is the appointment owner
+        if(appointment.user.toString() !== req.user.id && req.user.role !== 'admin'){
+            return res.status(401).json({
+                success: false,
+                message: `User ${req.user.id} is not authorized to update this appointment`
             });
         }
 
@@ -125,6 +149,15 @@ exports.deleteAppointment = async (req, res, next) => {
                 message: `No appointment with the id of ${req.params.id}`
             });
         }
+
+        //Make sure user is the appointment owner
+        if(appointment.user.toString() !== req.user.id && req.user.role !== 'admin'){
+            return res.status(401).json({
+                success: false,
+                message: `User ${req.user.id} is not authorized to delete this appointment`
+            });
+        }
+
         await appointment.deleteOne();
         res.status(200).json({
             success:true,
